bot/pyproject.toml
[project]
name = "bot"
version = "0.1.0"
dependencies = [
    "aiogram==3.25.0",
    "aiohttp==3.11.16",
    "asyncpg==0.31.0",
    "msgspec==0.20.0",
    "pymorphy3==2.0.3",
    "python-dotenv==1.0.1",
    "redis[hiredis]==7.1.1",
    "sqlalchemy==2.0.46",
    "tzdata==2025.3",
    "uvloop==0.22.1; sys_platform == 'linux' or sys_platform == 'darwin'",
]

[tool.uv]
package = false
